Refactor storage and comments. Finish TODO's

+++ src/dependencies/mpt/v2/Data.sol
@@ -3,12 +3,14 @@ pragma solidity >=0.5.0;
 library D {
     struct Label {
         bytes32 data;
-        uint length;
+        uint256 length;
     }
+
     struct Edge {
         bytes32 node;
         Label label;
     }
+
     struct Node {
         Edge[2] children;
     }
