+++ src/dependencies/mpt/v2/LibMPT.sol
@@ -1,19 +1,25 @@
 pragma solidity >=0.5.0;
 
-import { D }     from "src/dependencies/mpt/v2/Data.sol";
+import { D } from "src/dependencies/mpt/v2/Data.sol";
 import { Utils } from "src/dependencies/mpt/v2/Utils.sol";
 
 library LibMPT {
-    function verifyProof(bytes32 rootHash, bytes memory key, bytes memory value, uint branchMask, bytes32[] memory siblings) public pure {
+    function verifyProof(
+        bytes32 rootHash,
+        bytes memory key,
+        bytes memory value,
+        uint256 branchMask,
+        bytes32[] memory siblings
+    ) public pure {
         D.Label memory k = D.Label(keccak256(key), 256);
         D.Edge memory e;
         e.node = keccak256(value);
-        uint b = branchMask;
-        for (uint i = 0; b != 0; i++) {
-            uint bitSet = Utils.lowestBitSet(b);
-            b &= ~(uint(1) << bitSet);
+        uint256 b = branchMask;
+        for (uint256 i = 0; b != 0; i++) {
+            uint256 bitSet = Utils.lowestBitSet(b);
+            b &= ~(uint256(1) << bitSet);
             (k, e.label) = Utils.splitAt(k, 255 - bitSet);
-            uint bit;
+            uint256 bit;
             (bit, e.label) = Utils.chopFirstBit(e.label);
             bytes32[2] memory edgeHashes;
             edgeHashes[bit] = edgeHash(e);
